// src/lib/lembretes/types.ts
// Types for lembretes records
export type Lembrete = {
  id?: number; // Auto-generated by the system
  lembrete: string; // Filled by user in form
  tipo: string; // Filled by user in form
  valor?: number; // Filled by user in form
  telefone?: string; // Phone number from Clientes.telefone
  cliente?: string; // Name from Clientes.nome
  id_cliente?: string; // Auto-generated by the system
  vencimento: string; // Filled by user in form
  lembrar: string; // Same as vencimento
};